requirements.txt
click==8.3.1
colorama==0.4.6
gunicorn==25.3.0
h11==0.16.0
packaging==26.0
uvicorn==0.42.0
Django==6.0.3
djangorestframework==3.17.1
